==> Notebooks/05_transformations.ipynb <==
import sys
from pathlib import Path

# Ubicación del notebook
NOTEBOOK_DIR = Path.cwd()

# Raíz del proyecto = subir un nivel desde Notebooks/
PROJECT_ROOT = NOTEBOOK_DIR.parent

# Añadir raíz del proyecto al sys.path
if str(PROJECT_ROOT) not in sys.path:
    sys.path.insert(0, str(PROJECT_ROOT))

print("Proyecto raíz detectado:", PROJECT_ROOT)

"""Importamos las librerias"""

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from utils.paths import DATA_DIR, FIG_DIR
from utils.loader import load_csv

SERIES_DATA_DIR = DATA_DIR / "time_series"
SERIES_DATA_DIR.mkdir(parents=True, exist_ok=True)

print("📁 DATA_DIR:", DATA_DIR)
print("📁 FIG_DIR:", FIG_DIR)

"""Descargamos el dataset limpio"""

market_df_clean = load_csv(DATA_DIR / "clean_market_data.csv")


# --- Usamos directamente el DataFrame limpio ---
prices_df = market_df_clean.copy() # tu DataFrame ya normalizado, sin nulos, fechas alineadas

#Pivotamos el dataframe
prices_pivot = prices_df.pivot(index='date', columns='ticker', values='close')

# --- Selección del DataFrame base ---
prices = prices_pivot.copy()
prices.index = pd.to_datetime(prices.index, errors="coerce")
prices = prices.sort_index()

# --- Aplicamos transformaciones ---

# Logaritmo natural de los precios
log_prices = np.log(prices.replace({0: np.nan}))

# Dfierenciacion de precios
diff_close = prices.diff().dropna()

# Diferenciación logarítmica
log_diff = log_prices.diff().dropna()

# Retornos simples
returns = prices.pct_change().dropna()

# Retornos al cuadrado
returns_sq = returns**2

# Media móvil (21 días)
roll_mean_21 = prices.rolling(21).mean()

# Desviación estándar móvil (21 días)
roll_std_21 = prices.rolling(21).std()

"""Guardamos los dataframes transformados"""

prices.to_csv(SERIES_DATA_DIR / "prices.csv")
diff_close.to_csv(SERIES_DATA_DIR / "diff_close.csv")
log_prices.to_csv(SERIES_DATA_DIR / "log_prices.csv")
log_diff.to_csv(SERIES_DATA_DIR / "log_diff.csv")
returns.to_csv(SERIES_DATA_DIR / "returns.csv")
returns_sq.to_csv(SERIES_DATA_DIR / "returns_sq.csv")   
roll_mean_21.to_csv(SERIES_DATA_DIR / "roll_mean_21.csv")
roll_std_21.to_csv(SERIES_DATA_DIR / "roll_std_21.csv")
